handle grpc server not ready in OTLP exporter (#1211)
part of #1096 

This change fixes an issue where if the grpc server could not be reached
when the pipeline started, the OTLP exporter's start method would exit
early with no feedback. Now, we wait to try the connection until we need
it. This change adds test to ensure the exporter behaves as expected in
the case of the server not being ready when the pipeline starts and if
the server restarts (e.g. the connection is closed).

If this is OK as a solution, I'll add the same solution to OTAP exporter
as a followup