fix otlp exporter test flake (#1230)
Closes: https://github.com/open-telemetry/otel-arrow/issues/1231

Try to fix test failure observed here: 

https://github.com/open-telemetry/otel-arrow/actions/runs/18234929441/job/51926734884

The issue here was that, when we were asserting that the "logs_failed"
metric was incremented after we send a request which could not be
serviced, we'd send the pdata, then send the `CollectTelemetry` pdata
message, then wait to receive telemetry on the metrics reporter channel.
The problem with this is that the channel that gets fed into the
exporter biases control messages, so in some cases the exporter would
see the `CollectTelemetry` message first. When this happens, there could
be no telemetry to report, which means nothing would be sent on the
pdata reporter channel and the test would deadlock.

Because the test driver can't reliably control the order in which the
exporter sees these messages two messages in the failure case, we change
the test to just check at the of the pdata sequence and ensure we saw
the correct number of exported & failed messages. This is reliable
because before checking this, we send a pdata message that should
succeed and we can wait on the `req_receiver` ensure the client saw and
send this message before we send the `CollectTelemetry` message.

I ran the test in a loop 50x on this commit
https://github.com/open-telemetry/otel-arrow/pull/1230/commits/72e30d114daaa71e4918e53a486448cf3e1e426a
to ensure it passes reliably.