[query-engine] Relax restrictions in KQL summarize expressions (#1262)
## Changes

* Allow aggregate and group by expression without explicit assignment in
KQL summarize expressions

## Details

Today you can do:
* `summarize by Timestamp = bin(Timestamp, 1m)`
* `summarize Count = count()`

But you will get an error if you do:
* `summarize by bin(Timestamp, 1m)`
* `summarize count()`

This PR makes the second two examples work. In the first case the name
`Timestamp` will be used and in the second case `count`.

This is done by giving the KQL parser some logic to determine the
identifier for any given scalar expression. Note: We could probably use
this in extend\project* cases too but I plan to look at that in a
follow-up.

Fixes #1243

---------

Co-authored-by: Drew Relmas <drewrelmas@gmail.com>