[PerfTest] remove github actions from benchmarks branch (#1293)
https://github.com/open-telemetry/otel-arrow/issues/1292

The benchmarks branch holds results for benchmark suites and is used by
github-pages to build the sites page (from the docs/ directory).

We attempted to merge main back into this branch to fix failing CI
checks that prevent this branch from being updated via pull request in
https://github.com/open-telemetry/otel-arrow/pull/1285, but ran into
significant issues with the EasyCLA check.

Instead of going this route, we think it's better to strip this branch
of all non-essential files for its intended purpose. This is the first
of a few PRs:

1. This PR - strip workflow definitions to disable all automations on
this branch (the github-pages steps are defined elsewhere, none of these
workflows are useful on this branch).
2. Re-submit the changes for #1285 without the attempt to update to
main.
3. Further evaluate what files are important for this branch and remove
everything else.