[PerfTest] remove perf orchestrator v1 code and pipeline (#1320)
This PR removes the original pipeline perf orchestration components
which are unused in the new framework now active in CI jobs.

Specifically:

1. Remove the original CI job that was demonstrating the go
otel-collector with and without batches enabled.
2. Remove the original python backend (same functionality is supported
with a df_engine running perf-exporter for both otap and otlp).
3. Remove the original orchestrator.py script and associated library
files.
4. Remove the original (go collector) system under test config
directories - these are now located inside the various test_suites.
5. Remove the report_templates directory - these are now located inside
the various test_suites.
6. Remove the collector_batch_comparison and dataflow_engine_e2e_logs
test suites that were setup just to verify parity with the older
scripts.
7. Update readme files to reference the current state of the project.

> NOTE: The new framework does not yet have the support for kubernetes
that the old script had. Since this isn't in use anywhere in our current
pipelines, it seems safe to remove.