[query-engine] Fix up DateTime attributes in OTLP bridge when schema is present (#1336)
## Changes

* In OTLP bridge if schema says an attribute key should be a DateTime
but it is found to be a String attempt an automatic conversion

## Details

@drewrelmas has found users to be making multiple calls into the
recordset engine while processing the same batch of records. In the
first call a DateTime is created. In a later call that DateTime is used
in a `summary` (eg: `bin(TimeGenerated, 10m)`). The problem is, the
bridge input and output is OTLP where "extended" things like DateTimes
become Strings. This is a stopgap while we look at better solutions.