[PipelinePerf] setup verification of manual workflow runs for PRs (#1381)
These changes allow automated verification of a manually triggered perf
test workflow for a given PR number. Because of how github actions work
for this weird case, it's a 2 step thing:

- The new workflow file below (pipeline-perf-verify-pr.yaml) inserts a
placeholder status into PRs on create/update that will remain "pending"
indefinitely.
- The original manual workflow now updates that placeholder status with
the result of it's own run (for the specified PR number).

So the flow is:
- New PR created, new job inserts the pending status result
- Maintainer triggers the manual run with target PR number
- Manual job runs and sets the previously pending result to pass/fail

New commits will clear the result back to pending.