[PipelinePerf] replace marker + manual verification logic with labels (#1391)
Since the PR marker and manual workflow steps weren't working well, this
PR switches the logic to a single label-based workflow (the manual
workflow is left in place for executing specific suites):

When a new PR is created:
- Job 1 (on the standard ubuntu-latest runer) checks for the existence
of a label "pipelineperf" on the PR.
- If label is not found, Job 1 fails (so there's a better visual
indicator that it hasn't run, rather than the 'skipped' icon we see with
the cargo bench job).
- If the label is found, Job 1 sets an output that tells Job 2 to run
(on the dedicated hardware).

Updates to the PR once the label is applied will automatically run the
perf tests again, so we should avoid doing this until the PR is
basically ready to merge in order to avoid over-use of the dedicated
runner.