Move otel-arrow-rust into otap-dataflow (#1395)
Part of #867 
Fixes #1377

This is a move, no real code changes were required, only to replace
`otel_arrow_rust::` with `crate::` (in crates/pdata) or
`otap-df-pdata::` (in crates/otap). Top-level module files were merged
with the existing, so several files were manually combined, but there
were no name conflicts.

All relative paths change from otel-arrow-rust/${x} to
otap-dataflow/crates/pdata/${x}.

It is a pleasure to finally combine these two codebases. Thanks to
@v0y4g3r for the original code structure here. A lot has changed since
the first commit!

This crate had a `#[allow(missing_docs)]` declaration. I've removed it
from the crate-level and applied it in dozens of locations instead. I
will file a new issue reminding us that there are a lot of missing docs
in this crate; I think it will be easier to finish this documentation
now that it's more solidified.