Move OTAP encoding functions, OtapPayload struct & helpers to otap_df_pdata (#1402)
I am working on testing the OTAP batching functionality and I want to
use the encoding functions from otap_df_pdata, which is where they
belong.

The otap_df_otap::pdata::{Context, OtapPdata} types remain, but about
half of it moves to otap_df_pdata::payload.

Minor test cleanups, part of otap_df_otap::testing moves with it, some
of the same moves to the one location it was used.

#1377