fix repeated primitive fields using different encodings when converting between OTLP bytes & OTAP (#1421)
part of #1399 

This fixes an issue caused by a mismatch encodings used repeated
primitive proto fields. There are three fields of this type:
- HistogramDataPoint bucket_counts (repeated fixed64) & explicit_bounds
(repeated double)
- ExpHistogramDataPoint Bucket bucket_counts (repeated uint64)

When we encoded OTAP -> OTLP bytes, we were writing the "expanded"
encoding, but when decoding OTLP Bytes -> OTAP, we were expecting the
"packed" encoding (for more details on the encodings see
https://protobuf.dev/programming-guides/encoding/#repeated).

This corrects the encoding of OTAP -> OTLP bytes to use "packed", which
is the default since proto Edition 2023 & in proto3 (it is also what
prost uses).

In a followup PR, I will add correct the logic in our decoder to handle
the expanded decoding and additional edge cases for packed encoding.