Support expanded and packed encoding for repeated primitive fields when decoding protobuf (#1424)
closes #1399 

According to the [protobuf encoding
docs](https://protobuf.dev/programming-guides/encoding/), there's two
cases decoders need to handle:

> Protocol buffer parsers must be able to parse repeated fields that
were compiled as packed as if they were not packed, and vice versa.

Also, if the encoding is "packed", we need to handle if there are more
than one segment of the buffer containing packed values
> Note that although there’s usually no reason to encode more than one
key-value pair for a packed repeated field, parsers must be prepared to
accept multiple key-value pairs. In this case, the payloads should be
concatenated. Each pair must contain a whole number of elements. The
following is a valid encoding of the same message above that parsers
must accept

This PR brings our decoding logic in line with the recommendations for
the few fields we have that could have this encoding. Before this
change, we were assuming the encoding was always packed.

This change implements a new iterator for these types called
`RepeatedPrimitiveIter` which encapsulates the logic of determining how
to iterate the buffer and produce values correctly for both types of
encodings. Internally, it use a combination of the
`RepeatedFieldProtoBytesParser` and the packed field iterators which
were already implemented before this change.

---------

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>
Co-authored-by: Utkarsh Umesan Pillai <66651184+utpilla@users.noreply.github.com>