handle dict encoded attribute parent IDs when adding transport optimized encoding (#1446)
fixes #1442 

Fixes a bug in the code that adds the transport optimized encoding to
the parent_id column of attributes record batches where the parent_id is
a u32 type. For attributes, this column may be dictionary encoded which
was not being handled by the code which adds the transport optimized
encoding.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>
Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>