Support protobuf encoding default `AnyValue`s from missing optional columns (#1447)
fixes: #1445 

We have a handful of columns that that represent `AnyValue`'s value that
are optional columns. If the column is full of default values, we omit
these columns. However there was a bug in the encoding of these fields
into OTLP protobuf, we treated the missing column as a null value which
is not correct if the value_type column is not `ValueType::Empty`.

These values must be encoded in OTLP, despite their default value,
because they are protobuf `oneof` fields.