Initial API & structure of columnar query engine (#1453)
Closes #1417 & #1409 

During last night's SIG meeting for the query/transform engine working
group, there was some consensus that having our Pipelines made up of
multiple stages (instead of a single DataFusion `ExecutionPlan`), seemed
like a good option. This gives me confidence that the plan laid out in
#1417 is probably workable - so this PR implements what's spec'd out
there with some modifications based on feedback on the issue.

Adds the initial shell of the columnar query engine including the:
- `Pipeline` - top-level API for invoking an OPL/KQL pipeline on OTAP
Data
- `PlannedPipeline` - internal data structure used by the `Pipeline`
containing the stages and state needed for adapting to changing OTAP
batch schemas
- `PipelineStage` - trait implemented by a stage in the pipeline to
transform OTAP batch
- `DataFusionPipelineStage` - an implementation of `PipelineStage`
backed by a DataFusion `ExecutionPlan`
- `PipelinePlanner` - used to transform a the
[`PipelineExpression`](https://github.com/open-telemetry/otel-arrow/blob/a2fc3baf6988b2014a28f12006b56757c82ebc37/rust/experimental/query_engine/expressions/src/pipeline_expression.rs#L7)
into `PipelineStages` (not actually implemented)
- `RecordBatchPartitionStream` - implementation of
[`PartitionStream`](https://docs.rs/datafusion/latest/datafusion/physical_plan/streaming/trait.PartitionStream.html)
used to shuffle the current `RecordBatch` for a given OTAP Payload Type
into `DataFusionPiplineStage`'s `ExecutionPlan`

This gives us a concrete base which we can iterate on as we implement
the rest of the columnar engine.