Add github workflow for automatic Tidy on Renovate PRs (#1457)
Fixes #1456.

Adds a workflow to run `make genotelarrowcol` on top of any Renovate
dependency upgrades.

If it works properly, I thought it would be best to remove the bunching
together of all Golang upgrades together which currently exists in our
Renovate config. It will make it easier to quickly merge non-breaking
upgrades and isolating upgrades that require further manual
intervention.