Use otelbot app token for Tidy push (#1460)
Another attempt to get solution working for #1456.

It *almost* worked as expected in #1458 but I forgot an earlier lesson
from
[prepare-release](https://github.com/open-telemetry/otel-arrow/blob/382c61d4b35eaf8c4753b046d57aa7bc70d3b0e6/.github/workflows/prepare-release.yml#L133-L134)
that we have to inject the special otelbot app token as an environment
variable in the task that does the Git action. Otherwise, regular PR
workflows are not triggered.

I've seen some evidence that we should opt out of persisting regular git
creds from `checkout` in order for the `auth setup-git` to work. Would
like to try this configuration.