Batch processor refactoring, testing (#1462)
This is such an extensive change that we may as well call it a rewrite,
although the basic structure remains.

I have modified the Config slightly to match the original Collector
batch processor, with three fields (batch_size, max_batch_size,
timeout):

- Uses Option<NonZeroUsize> values instead of more-elaborate logic
around 0 values disabling features.
- Only one size is required, it can be either the lower or upper bound.
This component will only invoke the otap_df_pdata logic after it reaches
the smaller of the two sizes. The max-size (if set) is passed to the
lower-level utility for splitting.

I made the logic signal-generic; there is only one call to
send_message().

I included a fairly-large hack for timing: as the signals are
independent, and as the effect handler has no "one-shot" timer support,
I used the DelayData signal with an empty request of the signal that set
the timer. Previously the code was incomplete: the "TimerTick" message
was being used to flush, but no timers were set. Now, we set the timer
whenever a transition from 0-items happens; however we also reset this
and have no way to cancel, so this is a major TODO.

Modifies the internal metrics to match the new code structure: primary
focus is on producer/consumer items/batches so we can monitor
performance and item-count correctness.

Completely replaces the tests following the new DataGenerator; however
does not modify/extend/improve the test fixture support which remains
incomplete following #1428.

Adds a demo config.

---------

Co-authored-by: Utkarsh Umesan Pillai <66651184+utpilla@users.noreply.github.com>