Optimize function used to filter child record batches with u16 parent_ids (#1488)
part of #1390 

When benchmarking/profiling filtering for the columnar query engine
(#1477 / #1342, the poc), I noticed we were spending a lot of time in
`build_uint16_id_filter`.

This function is shared with the filter processor, and it is used after
we filter the parent record batch to remove rows in children that point
to rows that were filtered out. To do this, we take the set of IDs from
the parent, and then build a selection vector for the child record
batch.

This PR tries to optimize this function. I added a benchmark that swept
across:
- batch sizes
- number of child records per parent
- filter selectivities (e.g. what fraction of the IDs to filter)
- sort order
   - sorted by parent ID
- "transport optimized" sort order (where the parent IDs are mixed up
b/c we sorted by attribute keys/values first)

I then used the results of this benchmark to make the following
optmizations:

**1. Use the optimized `BooleanBuilder::append_n` when building up the
result.**
Turns out this is faster than just appending values one at a time:
https://github.com/apache/arrow-rs/blob/1c90efe717c1a2f99a0a97d6e9c93cd913eaca3c/arrow-buffer/src/builder/boolean.rs#L164-L196

**2. Use a bitmap for the ID set instead of a HashSet**
This saves us some cycles because we don't need to hash everything in
the ID column when doing lookups in the ID set.

**3. Be more restrictive when using the "simd" code path**
When the ID column is very long, but there are relatively few records in
the ID set, we determined that it was faster to use this algorithm that
invokes the arrow `eq` and `or_kleene` compute kernels. See the
discussion
[here](https://github.com/open-telemetry/otel-arrow/pull/1279#discussion_r2453094683).

There are some heuristics in place for when to choose this code path:
the ID column must be sufficiently long and the ID set must be short in
comparison.

However, if the ID column is very very long, the ID set could still be
short enough in comparison to the ID column, but still be so long that
we iterate this code path many times. This is slow!

I did some testing and figured out that, at least on my laptop, this
code path tends to be slower after about 20 iterations even if the ID
column is very very long. So now we check this as well.

---
We should probably do the same for the u32 version of this function,
which can happen in a followup.

---------

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>