Fix error when filtering logs if there are no attributes (#1489)
closes: #1484 

I noticed when testing some of the filtering code that if we have a Logs
OTAP batch with no attributes, an error will be thrown when filtering
b/c the ID column is not found. This column is optional if there are no
attributes. I corrected the log filtering code and added a test.

I also added a test for the trace filtering code for good measure, even
though it isn't afflicted with the same behaviour. It seems that for
traces, we treat the ID column as non-optional.