Add filtering benchmarks (#1499)
Adds benchmarks for the filtering code used in the filter processor and
for the columnar query engine, with similar filtering criteria.

My primary goal was to confirm that we get similar performance using the
both approaches. The filter processor code uses arrow's compute kernels,
whereas the columnar query engine uses a hybrid approach of datafusion's
physical expression and compute kernels.

**filter pattern** | **batch_size** | **pdata filter code (µs)** |
**col. query engine (µs)**
:--- | :---: | :---: | :---:
`severity_text == "WARN"` | 32 | 8.5488 | 7.6645
"  | 1024 | 37.295 | 22.341
" | 8192 | 233.7 | 153.21
`attrs["code.namespace"] == "main"` | 32 | 9.9075 | 10.202
" | 1024 | 52.496 | 43.148
" | 8192 | 408.4 | 331.72
`attrs["code.namespace"] == "main" or attrs["code.line.number"] == 2` |
32 | 11.666 | 13.387
" | 1024 | 70.035 | 63.779
" | 8192 | 572.73 | 510.23
`severity_text == "WARN" and attrs["code.namespace"] == "main"` | 32 |
11.705 | 10.719
" | 1024 | 48.919 | 35.641
" | 8192 | 377.17 | 274.38

The results leads me to believe the approach the columnar query engine
is taking for filtering is on the right track.

My secondary goal here was simply to get some benchmarks in place for
the columnar query engine's filtering code, so we have a baseline for
which to measure the effect of future performance optimizations and to
ensure we're not introducing regressions as we add more sophistication
to support additional filtering use cases.