columnar query engine filtering handle adaptive OTAP schemas (#1497)
closes #1494 

Adds code to the `FilterPipelineStage` for the columnar query engine to
handle all the interesting ways that OTAP batches might change change
from one batch to the next including:
- Presence of optional payload type's RecordBatch
- Presence of optional columns
- Column data types (dict / native)
- Column order rearranged in a given record batch

A new type called `AdaptivePhysicalExprExec` is added to encapsulate the
logic around all of this.

To handle optional `RecordBatch`s for some payloads,
`AdaptivePhysicalExprExec` lazily converts the logical `Expr` into the
`PhysicalExpr` only when a record batch for some type is present, which
means we don't need all payload types to be present in the first batch
we see.

To handle optional columns and columns changing order,
`AdaptivePhysicalExprExec` determine which columns are referenced in the
DataFusion logical `Expr` that defines the filter by visiting it using a
new type called`ProjectedSchemaExprVisitor`. The result of the
expression traversal is then captured in a type called
`FilterProjection`. Before applying the filter, we attempt do a
lightweight projection to create a new record batch with only the
referenced columns in the correct order. If the columns are missing,
this projection returns `None` and we treat it as no rows passed the
filter (will correct this logic in a future PR when we handle predicates
like `something == null`).

Tests are added to cover the new functionality and the tests that were
previously ignored in this module are now passing