Azure support for parquet exporter and cloud_auth module (#1517)
The primary motivation for this PR is adding support for azure backends
for the parquet exporter. This work seems potentially re-usable for
anything that interacts with object storage and/or anything that needs
an azure token credential, so I also did a bit of module shuffling.
Major changes:

- Added a `cloud_auth` module with support for an initial set of azure
token credential types
- Moved the `object_store` module up and added support for azure token
credential types
- Added a new `azure` feature flag that gates all things azure across
both `cloud_auth` and `object_store`
- Fixed up tests where appropriate

Something that I did not do at this stage was explore moving all of this
auth stuff into some kind of `identity` extension. That might be the
next step and I think much of the code here could be easily folded into
such an extension.

Related to: https://github.com/open-telemetry/otel-arrow/issues/501