Columnar query engine add short-circuit optimization to composite and/or filter execs (#1523)
closes #1506 

Adds an optimization to the columnar query engine's filtering code where
we can short-circuit the execution of and/or filters if depending on the
result of one side of the evaluation. For example:
- `x1 == y1 AND x2 == y2`, if `x1 == y1` matches zero rows, there is no
point in checking `x2 == y2`
- similarly `x1 == y1 OR x2 == y2`, if `x1 == y1` matches every row, ,
there is no point checking `x2 == y2`

This adds this optimization to `Composite<FilterExec>` for the `and` and
`or` case, and to `Composite<AttributeFilterExec>` for the `and` case
only.

The `or` case for attributes isn't done b/c we would need to check if
the predicate passes for every unique parent_id, which is so straight
forward so it's not clear if in some cases would introduce more overhead
than just executing the other side of the filter. Requires further
investigation.

Selected benchmark results showing the performance improvement of this
short-circuit optimzation:
```
and_attrs_short_circuit/batch_size/32
                        time:   [1.8433 µs 1.8522 µs 1.8600 µs]
                        change: [−49.152% −48.864% −48.549%] (p = 0.00 < 0.05)
                        Performance has improved.
and_attrs_short_circuit/batch_size/1024
                        time:   [4.5644 µs 4.5717 µs 4.5786 µs]
                        change: [−68.020% −67.941% −67.864%] (p = 0.00 < 0.05)
                        Performance has improved.
and_attrs_short_circuit/batch_size/8192
                        time:   [24.917 µs 24.956 µs 24.994 µs]
                        change: [−74.971% −74.895% −74.827%] (p = 0.00 < 0.05)
                        Performance has improved.

and_short_circuit/batch_size/32
                        time:   [824.82 ns 830.06 ns 835.07 ns]
                        change: [−71.132% −70.962% −70.795%] (p = 0.00 < 0.05)
                        Performance has improved.
and_short_circuit/batch_size/1024
                        time:   [1.3861 µs 1.3899 µs 1.3933 µs]
                        change: [−90.697% −90.664% −90.632%] (p = 0.00 < 0.05)
                        Performance has improved.
and_short_circuit/batch_size/8192
                        time:   [5.4423 µs 5.4488 µs 5.4551 µs]
                        change: [−95.432% −95.416% −95.401%] (p = 0.00 < 0.05)
                        Performance has improved.


or_short_circuit/batch_size/32
                        time:   [2.3195 µs 2.3265 µs 2.3330 µs]
                        change: [−44.904% −44.594% −44.270%] (p = 0.00 < 0.05)
                        Performance has improved.
or_short_circuit/batch_size/1024
                        time:   [23.992 µs 24.059 µs 24.124 µs]
                        change: [−18.687% −18.371% −18.042%] (p = 0.00 < 0.05)
                        Performance has improved.
or_short_circuit/batch_size/8192
                        time:   [127.00 µs 127.23 µs 127.49 µs]
                        change: [−22.200% −21.999% −21.791%] (p = 0.00 < 0.05)
                        Performance has improved.
  ```

As part of the implementation this PR also introduces an `IdMask` num type as the return value from `AttributeFilterExec::execute` method, with explicit variants for selecting `All`, `None`, `Some` IDs and a `NotSome` representing IDs not to select. The reasons for this are:
- it's a clear signal to the caller when `All` or `None` IDs should be selected (an empty or full RoaringBitmap could have been used, but [allocating a full roaring bitmap is expensive](https://docs.rs/roaring/latest/src/roaring/bitmap/inherent.rs.html#36))
- the `NotSome` variant allows us to  invert the selection of IDs without having to do a 2nd pass over the parent ID column (like we were doing [here](https://github.com/open-telemetry/otel-arrow/blob/f3080ac9bfc5673cf4fc8ac24776fd096f353546/rust/experimental/query_engine/engine-columnar/src/pipeline/filter.rs#L620-L626))

I will probably do a followup PR that tries to take better advantage of this `IdMask` to further optimize filter performance for attribute filtering.