Columnar query engine filter handle `== null` filters (#1538)
part of #1508 

Handles filter predicates like `severity_text == null` or
`attributes["x"] == null` (e.g. doesn't exist) when filtering using the
columnar query engine.

It handles all the cases, including when the optional column is not
present, when then the ID column (used to join attributes) is not
present or null (meaning no attributes), and when attribute batches are
entirely absent.

---------

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>