Add more options to df_engine docker build, slim down features for some packages (#1540)
This PR addresses a few problems that I ran into getting `df_engine`
running in my kubernetes environment. The end result of all this is that
you can now run a docker build like this: ` docker build --build-arg
FEATURES="azure" --build-arg RUSTFLAGS="-C target-feature=-gfni"
--build-arg DEBUG=1 --build-context otel-arrow=../../ -f Dockerfile -t
df_engine .`

# TLS

Some of the Azure SDK crates + reqwest have a lot of default features
enabled that we are not using and which rely on linking with openssl for
TLS. This was giving me trouble with our musl build.

I've disabled default features for these packages at the workspace level
and opted for using `rustls-tls` with reqwest in the `otap` crate.

# Configuring the docker build

- Added the `FEATURES` build arg to the dockerfile and the cross
platform build script so that you can turn on features like `azure` when
building
- Added the ability to specify `RUSTFLAGS` for the docker build so that
we can turn some features on or off when compiling which brings me to my
next item...

# Adding some optional debug utils to the image

When running on k8s I was getting very mysterious crashes after a couple
of seconds with 0 logs to stdout. Remoting into the container and
running the binary gave me the "Illegal instruction (core dumped)"
message, which was a start.

After checking the obvious vector extensions like SSE/AVX/AVX512, I was
stumped. So, I added gdb to the container and found the exact
instruction which is the longest mnemonic I have ever seen (comes from
arrow's buffer crate):

```
(gdb)  x/i $pc
=> 0x7ffff745fb4f <new+463>:    vgf2p8affineqb $0x0,%ymm3,%ymm5,%ymm5
```

Turns out my laptop has support for a specific feature called `gfni`,
which the nodes with Intel Xeon processors that I was running did not.

I think this will not be the last time someone needs to debug a native
code/cross-compilation problem, so I added an option to the build called
`DEBUG` that will include `gdb` (and maybe other debug utilities in the
future) in the image.