columnar query engine filtering: add tests for value is not null (#1551)
part of #1508 

In https://github.com/open-telemetry/otel-arrow/pull/1538 we added the
capability to filter where some property is null. This PR adds test
coverage for the inverese, e.g. `severity_text != null` or
`attributes["x"] != null`.

There are no code changes, this just adds unit tests. I wanted to have
coverage over this to avoid regressions as we continue to work on this
filtering code.