move columnar query engine into otap-dataflow (#1580)
Moves the columnar query engine from experimental into otap-dataflow.

The primary reason I'd like to move this is for other projects that
might depend on otap-dataflow pdata crate. For example, if a crate were
to depend on otap-dataflow and locate it through git, like:
```
otap-df-pdata = { git = "https://github.com/open-telemetry/otel-arrow.git", package = "otap-df-pdata", #...
```
Cargo will try to clone the git repo and scan through it to find a crate
called `otap-df-pdata`. However, cargo may encounter
query_engine/engine-columnar/Cargo.toml which depends on `otap-df-pdata`
by a relative path, and get confused, resulting in an error like:
```
error: no matching package named `otap-df-pdata` found
```
In summary, I think this probably would have needed to happen anyway
before we publish on crates.io
https://github.com/open-telemetry/otel-arrow/issues/1340

If the reason above is not convincing, we can additionally justify
moving this into otap-dataflow on the following merits:
1. the implementation is quite specific to OTAP, in that the pipeline
stages operate on and produce `OtapArrowRecords`
2. the experimental stage of this implementation was acheived in the POC
https://github.com/open-telemetry/otel-arrow/pull/1342, and currently
we're trying to build a columnar query engine that is of comparable
quality/maturity with the other otap-dataflow crates.

I realize the README and rust docs are currently light on content, but I
will happily update those in a followup PR if we're OK with relocating
this code.