Add perf test to saturate the df engine (#1612)
Trying to work towards
https://github.com/open-telemetry/otel-arrow/issues/1531
This adds a continuous tests, matching the existing back pressure ones.
There is only one difference - the load generator is not capped at 100K
RPS - it goes as much as it can go.


Need to try tweaking few settings to get it right. The engine and
load-generator is now running in just 1 core, we might need to run
load_generator on more than one core to fully saturate the engine..
Also, in future, we want to run engine of multiple cores to see if we
truly scale with number of cores..

Add as continuous to begin with, once things are stabilized, we can move
this to nightly runs.