columnar query engine support filtering by datetimes and additional int types (#1627)
part of #1508 

Adds support for filtering by integer types other than int64 and support
for filtering by timestamp to the columnar query engine.

Before this change, every int field we encountered we assumed was int64,
which worked when filtering by attribute values. However, there are a
handful of fields in the OTel/OTAP data model that are not of this type
so now we must check if we're applying a filter to one of these fields
and, if so, create a literal for the binary expression with the correct
integer type.

This PR also adds support for datetime literal - for example:
```kql
logs | where time_unix_nano > datetime(1970-01-01 00:00:01.1)
```