PerfResults - use PR title only in charts (#1651)
<img width="1158" height="514" alt="image"
src="https://github.com/user-attachments/assets/871e3cf9-6077-47bb-8c07-df58e0fcfe24"
/>
Since we (we here refers to GH Actions Benchmarks) use PR Title + PR
Description of each commit, it's quite hard to read the actual value
when viewing the charts.

This PR modifies to just use the PR Title. There is no built-in way to
do this, so had to do split on "\n". While at it, I also moved the logic
to a shared script as well.