Fix null handling in columnar query engine filtering code (#1688)
closes #1680 

Fixes the handling of null values in our filtering code. We expect the
selection vectors we combine for each filter criteria not to contain
nulls. This corrects the behaviour.

When comparing values using binary operators like
[`eq`](https://arrow.apache.org/rust/arrow/compute/kernels/cmp/fn.eq.html),
arrow's compute kernels will produce `null` in the resulting boolean
array. `nulls` cause problems because the `nulls` would propagate when
we combine the selection vectors using `and`, `or` and `not`. For
example, arrow's
[`or`](https://arrow.apache.org/rust/arrow/compute/fn.or.html) treats
`true | null` as null.

We'd like to treat expect `null == "val"` as `false` and we expect to
produce selection vectors that have no nulls. The solution is to combine
the null buffer that results from executing the physical expression with
the values buffer, to remove the nulls.