Fix bugs transforming attributes when all attrs deleted (#1687)
closes #1681 

- Fixes a bug where calling `transform_attributes` would produce an
error when attribute keys are dictionary encoded and all attributes are
deleted.
- Adds proper handling of the situation where this is called (in
`AttributesProcessor` and `AttributesTransformFilterStage`), which is
that the `RecordBatch` for the attribute's payload type should be
removed from the OTAP batch.
- Consolidates the transform attributes into a single function. Before
this, we had two functions `transform_attributes` and
`transform_attributes_with_stats` that had almost identical logic (the
differences just being stats returned, and comments)
- Fixes the to_string method on various `otap_df_pdata::error::Error`
variants that wrapped `ArrowError` but didn't expose the underlying
error in their error message.