Columnar query engine support conditionally executing pipeline stages (#1684)
part of #1667 

Adds a new `ConditionalDataExpression` to the transformation expression
AST for applying transformation `PipelineStages` to some subset of rows
that match a condition. This is used to implement a
`ConditionalPipelineStage`, which operates like `if/else if/else`
control flow.

For example, imagine we had a hypothetical syntax like:
```kql
logs |
  if (severity_text == "ERROR") {
     set attributes["important"] = "very" | set attributes["triggers_alarm"] = true
  } else if (severity_text == "WARN) {
     set attributes["important"] = "somewhat"
  } else {
     set attributes["important"] = "no"
  }
```

This could be modeled using our conditional expression like:
```rs
// this is pesudocode to illustrate what each field represents
Conditional {
  branches: [
     ConditionalBranch {
       condition: "severity_text == \"ERROR\"",
       expressions: [ 
         "set attributes[\"important\"] = \"very\"",
         "set attributes[\"triggers_alarm\"] = true"  
      ],
     },
     ConditionalBranch {
       condition: "severity_text == \"WARN\"",
       expressions: [
        "set attributes[\"important\"] = \"somewhat\"
      ],
     },
  ],
  default_branch: Some([
    "set attributes["important"] = \"no\""
  ])
}
```

Note there is currently no parser support for a language syntax that
creates this variant of `DataExpression`. That will happen in a future
PR

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>